ENH: Use a temporary connection file on the remote server
When opening a connection, store the connection info in a temporary file on the server (`~/remote_kernel.json`).
When the kernel exits, or a CTRl-C event is received locally, this file is removed.
This is a more secure option than supplying the connection arguments on the command line, as these arguments are visible in the list of processes on the server, allowing other user to connect to this kernel.

Add optional argument --no-remote-files to use the type connection, with the connection info passed on the command line.