ENH: Rename argument, check specification version
To prevent conflicts between kernel definitions from before and after recent changes, explicitly name the pre-command argument as `-pc` or `--pre-command`.
This also allows checking for version of kernel definitions, as older versions will have the deprecated `-c` or `--command` argument. If such a deprecated kernel specification is found, convert it to the newer version and re-save the specification file.